Synthetic commit for incomplete tag ctj-4-5-post
svn:revision:15815
cvs:account:cvs-fast-export

+++ util/loop.c
@@ -22,13 +22,8 @@
 static char RCSid[] = "@(#)$Header$ (ARL)";
 #endif
 
-#include "conf.h"
-
 #include <stdio.h>
 #include <math.h>
-
-#include "machine.h"
-#include "externs.h"
 
 #define	INTEGER 0
 #define	REAL	1
@@ -52,7 +47,7 @@ char	**argv;
 
 	/* determine if any arguments are real */
 	for (i = 1; i < argc; i++) {
-		if (atof(argv[i]) != ((double)atoi(argv[i]))) {	
+		if (atof(argv[i]) != atoi(argv[i])) {	
 			status = REAL;
 			break;
 		}
